requirements.txt
# Core CAMEL-AI
camel-ai>=0.2.0

# LLM Backend (OpenAI-compatible for Llama 3.3 via NVIDIA NIM / Groq / Together)
openai>=1.0.0

# NeMo Guardrails
nemoguardrails>=0.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
tenacity>=8.0.0
